kymamodule:  Implement resource tracking (#373)
* kymamodule: Add finalizer, implement resource tracking

* kymamodule: fix finalizer issue.

* kymamodule: remove unnecessary finalizer useage

* kymaModuletest: Update and add tests

* mykamoduletest - add test for checking correct parameter

* kymamodule: Remove redundant checks

* kymamodule fix: Add client check

* fix: Removed track call from Observe()

* kymamodule fix: handle error

* add tests for connect

* kymamodule: fix lint error

+++ internal/controller/kymamodule/kymamodule.go
@@ -2,6 +2,7 @@ package kymamodule
 
 import (
 	"context"
+	"fmt"
 
 	xpv1 "github.com/crossplane/crossplane-runtime/apis/common/v1"
 
@@ -15,14 +16,16 @@ import (
 	"github.com/sap/crossplane-provider-btp/apis/environment/v1alpha1"
 	"github.com/sap/crossplane-provider-btp/internal/clients/kymamodule"
 	"github.com/sap/crossplane-provider-btp/internal/tracking"
+	kerrors "k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/types"
 )
 
 var (
-	errNotKymaModule   = "managed resource is not a KymaModule custom resource"
-	errTrackPCUsage    = "cannot track ProviderConfig usage"
-	errTrackRUsage     = "cannot track ResourceUsage"
-	errSetupClient     = "cannot setup KymaModule client"
-	errObserveResource = "cannot observe KymaModule"
+	errNotKymaModule                  = "managed resource is not a KymaModule custom resource"
+	errTrackRUsage                    = "cannot track ResourceUsage"
+	errSetupClient                    = "cannot setup KymaModule client"
+	errObserveResource                = "cannot observe KymaModule"
+	errKymaEnvironmentBindingNotFound = "cannot get referenced KymaEnvironmentBinding"
 )
 
 // A connector is expected to produce an ExternalClient when its Connect method
@@ -44,15 +47,11 @@ type external struct {
 	secretfetcher SecretFetcherInterface
 }
 
-// This methods connects to the Kyma cluster using the kubeconfig from the KymaEnvironmentBinding
+// Connect connects to the Kyma cluster using the kubeconfig from the KymaEnvironmentBinding
 func (c *connector) Connect(ctx context.Context, mg resource.Managed) (managed.ExternalClient, error) {
 	cr, ok := mg.(*v1alpha1.KymaModule)
 	if !ok {
 		return nil, errors.New(errNotKymaModule)
-	}
-
-	if err := c.usage.Track(ctx, mg); err != nil {
-		return nil, errors.Wrap(err, errTrackPCUsage)
 	}
 
 	if err := c.resourcetracker.Track(ctx, mg); err != nil {
@@ -69,6 +68,9 @@ func (c *connector) Connect(ctx context.Context, mg resource.Managed) (managed.E
 	}
 
 	svc, err := c.newServiceFn(creds)
+	if err != nil {
+		return nil, errors.Wrap(err, errSetupClient)
+	}
 
 	return &external{
 			client:        svc,
@@ -76,7 +78,7 @@ func (c *connector) Connect(ctx context.Context, mg resource.Managed) (managed.E
 			kube:          c.kube,
 			secretfetcher: secretfetcher,
 		},
-		err
+		nil
 }
 
 func (c *external) Observe(ctx context.Context, mg resource.Managed) (managed.ExternalObservation, error) {
@@ -84,6 +86,45 @@ func (c *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 	cr, ok := mg.(*v1alpha1.KymaModule)
 	if !ok {
 		return managed.ExternalObservation{}, errors.New(errNotKymaModule)
+	}
+
+	// Get the reference for the kyma environment binding
+	if cr.Spec.KymaEnvironmentBindingRef == nil {
+		cr.SetConditions(xpv1.Unavailable().WithMessage(
+			"KymaEnvironmentBindingRef must be specified",
+		))
+		return managed.ExternalObservation{ResourceExists: true, ResourceUpToDate: false}, nil
+	}
+	binding := &v1alpha1.KymaEnvironmentBinding{}
+	bindingName := types.NamespacedName{
+		Name:      cr.Spec.KymaEnvironmentBindingRef.Name,
+		Namespace: cr.GetNamespace(),
+	}
+
+	if err := c.kube.Get(ctx, bindingName, binding); err != nil {
+		if kerrors.IsNotFound(err) {
+			// Reference is gone - can't be processed
+			cr.SetConditions(xpv1.Unavailable().WithMessage(fmt.Sprintf("Referenced KymaEnvironmentBinding %s doesn't exist", binding.Name)))
+			return managed.ExternalObservation{ResourceExists: false}, nil
+		}
+		return managed.ExternalObservation{}, errors.Wrap(err, errKymaEnvironmentBindingNotFound)
+	}
+
+	// Check if the binding is being deleted
+	if binding.GetDeletionTimestamp() != nil {
+		cr.SetConditions(xpv1.ReconcileSuccess().WithMessage(
+			"KymaEnvironmentBinding is pending deletion. Delete this KymaModule to allow binding deletion.",
+		))
+	}
+
+	if c.client == nil {
+		cr.SetConditions(xpv1.Unavailable().WithMessage(
+			"Cannot connect to Kyma cluster - kubeconfig may be unavailable or expired",
+		))
+		return managed.ExternalObservation{
+			ResourceExists:   true,
+			ResourceUpToDate: false,
+		}, nil
 	}
 
 	res, err := c.client.ObserveModule(ctx, cr)
@@ -106,7 +147,7 @@ func (c *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 }
 
 // Disconnect is a no-op for the external client to close its connection.
-// Since we dont need this, we only have it to fullfil the interface.
+// Since we dont need this, we only have it to fulfill the interface.
 func (c *external) Disconnect(ctx context.Context) error {
 	return nil
 }
@@ -130,7 +171,6 @@ func (c *external) Create(ctx context.Context, mg resource.Managed) (managed.Ext
 }
 
 func (c *external) Update(ctx context.Context, mg resource.Managed) (managed.ExternalUpdate, error) {
-
 	return managed.ExternalUpdate{}, errors.New("Update is not implemented - should not be called, only create")
 }
 
@@ -144,7 +184,16 @@ func (c *external) Delete(ctx context.Context, mg resource.Managed) (managed.Ext
 		return managed.ExternalDelete{}, nil
 	}
 
-	err := c.client.DeleteModule(ctx, cr.Spec.ForProvider.Name)
+	// Delete the module from Kyma cluster
+	if c.client != nil {
+		err := c.client.DeleteModule(ctx, cr.Spec.ForProvider.Name)
+		if err != nil {
+			return managed.ExternalDelete{}, err
+		}
+	}
 
-	return managed.ExternalDelete{}, err
+	// ResourceUsage cleanup is handled automatically by Kubernetes garbage collection
+	// No manual untracking or finalizer cleanup needed
+
+	return managed.ExternalDelete{}, nil
 }
